+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.2.7 2026
 
+* [Improve synchronization of plugin lifecycle and shortcut key declarations](https://github.com/siyuan-note/petal/issues/59)
 * [Add plugin command execution context](https://github.com/siyuan-note/siyuan/issues/19076)
 * [Add Protyle tracked range APIs](https://github.com/siyuan-note/siyuan/issues/19054)
 * [Add dynamic editor toolbar item plugin APIs](https://github.com/siyuan-note/siyuan/issues/19083)