+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v1.2.7 2026
+## v1.2.7 2026-08
 
 * [Improve synchronization of plugin lifecycle and shortcut key declarations](https://github.com/siyuan-note/petal/issues/59)
 * [Add plugin command execution context](https://github.com/siyuan-note/siyuan/issues/19076)